framework update

+++ js/controllers.js
@@ -10,6 +10,7 @@ angular.module('myApp.controllers', []).
 
     }])
     .controller('AboutCtrl', ['$rootScope', function ($rootScope) {
+        $rootScope.appDisplayName = "About";
         $rootScope.showlogin = false;
     }])
 .controller('ReportIssueCtrl', ['$scope', '$rootScope', '$http', '$location',
@@ -17,9 +18,9 @@ angular.module('myApp.controllers', []).
                                 if (window.device) {
                                 var userID="";
                                 if($.jStorage.get('username')==null || $.jStorage.get('username') =="")
-                                {  userID= "Not available";}
+                                {  userID= "";}
                                 else{
-                                userID = $.jStorage.get('username');
+                                userID = 'User ID : ' +$.jStorage.get('username');
                                 }
                                 window.plugin.email.isServiceAvailable(
                                                                        function (isAvailable) {
@@ -34,7 +35,7 @@ angular.module('myApp.controllers', []).
                                                                                                                 'Device version : ' + window.device.version + '<br/>' +
                                                                                                                 'Device model : ' + window.device.model + '<br/>' +
                                                                                                                 'Build Version : ' + $rootScope.metadata.version + '<br/>' +
-                                                                                                                'User ID : ' + userID + '<br/>' +
+                                                                                                                userID + '<br/>' +
                                                                                                                 '</p>',
                                                                                                                 isHtml: true
                                                                                                                 });}
@@ -54,9 +55,9 @@ angular.module('myApp.controllers', []).
                                  if (window.device) {
                                  var userID="";
                                  if($.jStorage.get('username')==null || $.jStorage.get('username') =="")
-                                 {  userID= "Not available";}
+                                 {  userID= "";}
                                  else{
-                                 userID = $.jStorage.get('username');
+                                 userID = 'User ID : ' +$.jStorage.get('username');
                                  }
                                  window.plugin.email.isServiceAvailable(
                                                                         function (isAvailable) {
@@ -70,7 +71,7 @@ angular.module('myApp.controllers', []).
                                                                                                                  'Device version : ' + window.device.version + '<br/>' +
                                                                                                                  'Device model : ' + window.device.model + '<br/>' +
                                                                                                                  'Build Version : ' + $rootScope.metadata.version + '<br/>' +
-                                                                                                                 'User ID : ' + userID + '<br/>' +
+                                                                                                                 userID + '<br/>' +
                                                                                                                  '</p>',
                                                                                                                  isHtml: true
                                                                                                                  });}
@@ -86,6 +87,8 @@ angular.module('myApp.controllers', []).
                                  }])
 
     .controller('DeviceCtrl', ['$scope', '$rootScope', function ($scope, $rootScope) {
+        
+                               $rootScope.appDisplayName = "Device Info";
         if(window.device) {
             $scope.devicename = window.device.name;
             $scope.deviceuuid = window.device.uuid;
@@ -104,6 +107,7 @@ angular.module('myApp.controllers', []).
     }])
     .controller('LoginCtrl', ['$scope', '$rootScope', '$http', '$location',
         function ($scope, $rootScope, $http, $location) {
+                              MyCampusApp.homeScreenDisplayed=false;
             $rootScope.back = function () {
                 $location.path("/home");
             };
@@ -119,6 +123,7 @@ angular.module('myApp.controllers', []).
                 };
                 var iabClose = function(data) {
                     window.location.href = "index.html#login";
+                              $.unblockUI();
                 };
                 iabRef.addEventListener("exit", iabClose);
                 var loadStop = function(data) {
@@ -169,7 +174,7 @@ angular.module('myApp.controllers', []).
                     }
                 }, function (data) {
                     //alert("Error call back" + data);
-                    apprise("Error occured while processing this request.", {'verify': false, 'textYes': "Ok"}, function (r) {
+                    apprise("Error occurred while processing this request.", {'verify': false, 'textYes': "Ok"}, function (r) {
                         $rootScope.ticket = null;
                         $rootScope.loggedin = false;
                         $rootScope.userroles = null;
@@ -187,6 +192,7 @@ angular.module('myApp.controllers', []).
     .controller('SettingsCtrl', ['$rootScope', '$scope', '$http', '$location', '$window', '$sce', '$route', '$compile',
         function ($rootScope, $scope, $http, $location, $window, $sce, $route, $compile) {
             //$scope.serverUrl = "http://localhost:8081/";
+            $rootScope.appDisplayName = "Settings";
             $rootScope.showlogin = false;
             $scope.serverUrlChange = function () {
                 $rootScope.serverUrl = $scope.serverUrl;
@@ -221,23 +227,53 @@ angular.module('myApp.controllers', []).
                 $('#listbtn').attr('class', 'btn btn-befault');
             }
 
-            $scope.resetmyapp = function () {
-                var tenant = MyCampusApp.config.tenant;
-                $.jStorage.deleteKey(tenant + '-metadata');
-                $location.path("/home");
-            }
+                                 $scope.resetApp = function(btn) {
+                                 if (btn == 1) {
+                                 var tenant = MyCampusApp.config.tenant;
+                                 $.jStorage.deleteKey(tenant + '-metadata');
+                                 $location.path("/home");
+                                 $scope.$apply();
+                                 }
+                                 }
+                                 $scope.resetmyapp = function() {
+                                 navigator.notification.confirm("Resetting your app clears all information associated with your app and initializes it with the default setting.  You can use this feature if your app is not functioning correctly. This feature requires network connectivity.", $scope.resetApp, 'Reset App', ['Yes', 'No']);
+                                 }
         }])
     .controller('HomeCtrl', ['$rootScope', '$scope', '$http', '$location', '$window', '$sce', '$route', '$compile',
         function ($rootScope, $scope, $http, $location, $window, $sce, $route, $compile) {
             //alert("Home controller called..");
             //alert ("Home Controller called");
             //setTimeout(function () {
+                             
+                             
+                             if(!$rootScope.hamburgerEvent)
+                             {
+                                $rootScope.$on("mobile-angular-ui.toggle.toggled", function(event, data) {
+                                            //alert(JSON.stringify(data));
+                                            if(data == "mainSidebar"){
+                                            if($rootScope.blackOverlay == undefined || $rootScope.blackOverlay == null){
+                                            $rootScope.blackOverlay=false;
+                                            }
+                                            if($rootScope.blackOverlay == false){
+                                            $rootScope.blackOverlay=true;
+                                            }else{
+                                            $rootScope.blackOverlay=false;
+                                            }
+                                            $rootScope.$apply();
+                                             }
+                                            
+                                            });
+                             
+                                $rootScope.hamburgerEvent=true;
+                             }
+                             
 
-
+                            
+                             
             if(!$rootScope.homeDownloadCompleteAdded) {
                 $rootScope.$on("onDownloadComplete", function(event, data) {
                     $.unblockUI();
-                    $route.reload();
+                    //$route.reload();
                     $rootScope.$apply(function () {
                         $location.path("/home");
                     });
@@ -438,7 +474,7 @@ angular.module('myApp.controllers', []).
                                 $.jStorage.deleteKey('userroles');
                             };
                             //handler();
-                            //navigator.notification.alert("Error occured while processing this request.", handler, 'Authentication', 'Ok');
+                            //navigator.notification.alert("Error occurred while processing this request.", handler, 'Authentication', 'Ok');
                         });
                     }catch(e) {
                         //ignore the exception as its silent authentication
@@ -449,6 +485,16 @@ angular.module('myApp.controllers', []).
     .controller('AppCtrl', ['$scope', '$routeParams', '$compile', '$http', '$rootScope', '$sce', '$window',
         '$location',
         function ($scope, $routeParams, $compile, $http, $rootScope, $sce, $window, $location) {
+            $rootScope.myswiperight = function() {    
+                if (!$("body").hasClass("sidebar-left-in")) {
+                    $rootScope.toggle("mainSidebar");
+                }                          
+            }
+            $rootScope.myswipeleft = function() {
+                if ($("body").hasClass("sidebar-left-in")) {
+                    $rootScope.toggle("mainSidebar");
+                }                
+            }
             MyCampusApp.homeScreenDisplayed = false;
             $scope.appname = $routeParams.appid;
             $scope.pageid = $routeParams.pageid;
@@ -527,7 +573,7 @@ angular.module('myApp.controllers', []).
             } catch (e) {
 
                 $.unblockUI();
-                apprise("Unknown error occured while processing the request.!", {'verify': false, 'textYes': "Ok"}, function (r) {
+                apprise("Unknown error occurred while processing the request.", {'verify': false, 'textYes': "Ok"}, function (r) {
                     $rootScope.back();
                 });
             }
@@ -581,9 +627,10 @@ angular.module('myApp.controllers', []).
 			$rootScope.appDisplayName = "KRYPTOS AppMaker";
 			var url = "https://kryptos.kryptosmobile.com";
 			$scope.loadApps = function(token) {
-				var message = '<div style="margin: 2px; vertical-align: middle; display: inline-block"><i class="icon-cog icon-spin icon-4x"></i><h3 style="color:white;">Loading ..!!</h3></div>';
-				$.blockUI({message : message});
-
+				//var message = '<div style="margin: 2px; vertical-align: middle; display: inline-block"><i class="icon-cog icon-spin icon-4x"></i><h3 style="color:white;">Loading ..!!</h3></div>';
+				//$.blockUI({message : message});
+                                   $.blockUI();
+                                   
 				$http.get(url + "/api/mobapp/listMyApps?token=" + token).
 					success(function (dt) {
 						//alert ("List Apps data : " + d1);
@@ -650,9 +697,9 @@ angular.module('myApp.controllers', []).
 					    errorCB, 'No Network', 'Ok');
 					return;
 				}
-				var message = '<div style="margin: 2px; vertical-align: middle; display: inline-block"><i class="icon-cog icon-spin icon-4x"></i><h3 style="color:white;">Authenticating</h3></div>';
-				$.blockUI({message : message});
-
+				//var message = '<div style="margin: 2px; vertical-align: middle; display: inline-block"><i class="icon-cog icon-spin icon-4x"></i><h3 style="color:white;">Authenticating</h3></div>';
+				//$.blockUI({message : message});
+                                   $.blockUI();
 				$http.post(url + "/api/authenticate" , {username : username, password : password}).
 					success(function(data) {
 						$rootScope.klogin = true;
@@ -717,9 +764,9 @@ angular.module('myApp.controllers', []).
 					return;
 				}
 
-			var message = '<div style="margin: 2px; vertical-align: middle; display: inline-block"><i class="icon-cog icon-spin icon-4x"></i><h3 style="color:white;">Loading the App..</h3></div>';
-			$.blockUI({message : message});
-
+			//var message = '<div style="margin: 2px; vertical-align: middle; display: inline-block"><i class="icon-cog icon-spin icon-4x"></i><h3 style="color:white;">Loading the App..</h3></div>';
+			//$.blockUI({message : message});
+                               $.blockUI();
 				var baseUrl = MyCampusApp.config.serverUrl;
 				$.jStorage.deleteKey(tenant + '-metadata');
 				var processLogosDir = function (logosDir) {
@@ -758,9 +805,9 @@ angular.module('myApp.controllers', []).
 			$rootScope.kappid;
 			var url = "https://kryptos.kryptosmobile.com";
 			$scope.loadApplications = function(token) {
-				var message = '<div style="margin: 2px; vertical-align: middle; display: inline-block"><i class="icon-cog icon-spin icon-4x"></i><h3 style="color:white;">Loading App..!!</h3></div>';
-				$.blockUI({message : message});
-
+				//var message = '<div style="margin: 2px; vertical-align: middle; display: inline-block"><i class="icon-cog icon-spin icon-4x"></i><h3 style="color:white;">Loading App..!!</h3></div>';
+				//$.blockUI({message : message});
+                                     $.blockUI();
 				$http.get(url + "/api/mobapp/appFeatures?id=" + $rootScope.kappid + "&token=" + token).
 					success(function (d1) {
 						$scope.appFeatures = d1;
